import-python-server/requirements.txt
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
coverage==7.13.2
Flask==3.1.2
flask-cors==6.0.2
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg-pool==3.3.0
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.0.0
python-dotenv==1.2.1
python-ulid==3.1.0
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
